extract table element to ProjectResults.vue
